+++ src/zoo-modules/form-modules/checkbox-module/checkbox.js
@@ -1,4 +1,4 @@
-import FormElement from '../common/FormElement.js';
+import { FormElement } from '../common/FormElement.js';
 
 /**
  * @injectHTML
@@ -6,53 +6,24 @@ import FormElement from '../common/FormElement.js';
 export class Checkbox extends FormElement {
 	constructor() {
 		super();
-	}
-
-	// TODO think of a better way to handle disabled attribute change
-	// TODO move checkbox inside label so handle multiple checkboxes
-	mutationCallback(mutationsList) {
-		for (let mutation of mutationsList) {
-			if (mutation.type === 'attributes') {
-				if (mutation.attributeName == 'disabled') {
-					if (mutation.target.disabled) {
-						this.setAttribute('disabled', '');
-					} else {
-						this.removeAttribute('disabled');
-					}
-				}
+		this.observer = new MutationObserver(mutationsList => {
+			for (let mutation of mutationsList) {
+				mutation.target.disabled ? this.setAttribute('disabled', '') : this.removeAttribute('disabled');
 			}
-		}
-	}
-
-	connectedCallback() {
-		const checkboxSlot = this.shadowRoot.querySelector('slot[name="checkbox"]');
-		checkboxSlot.addEventListener('slotchange', () => {
-			this.observer = this.observer || new MutationObserver(this.mutationCallback.bind(this));
-			let checkbox = checkboxSlot.assignedElements()[0];
-			checkbox.addEventListener('change', () => this.handleChange(checkbox));
+		});
+		this.shadowRoot.querySelector('slot[name="checkbox"]').addEventListener('slotchange', e => {
+			let checkbox = [...e.target.assignedElements()].find(el => el.tagName === 'INPUT');
+			if (!checkbox) return;
+			checkbox.addEventListener('change', () => this.toggleAttribute('checked'));
 			this.registerElementForValidation(checkbox);
-			if (checkbox.hasAttribute('disabled')) this.setAttribute('disabled', '');
-			this.observer.disconnect();
-			this.observer.observe(checkbox, { attributes: true, childList: false, subtree: false });
-			this.handleChange(checkbox);
+			if (checkbox.disabled) this.setAttribute('disabled', '');
+			if (checkbox.checked) this.setAttribute('checked', '');
+			this.observer.observe(checkbox, { attributes: true, attributeFilter: ['disabled'] });
 		});
 	}
 
-	handleChange(checkbox) {
-		if (checkbox.checked) {
-			checkbox.setAttribute('checked', '');
-			this.setAttribute('checked', '');
-		} else {
-			checkbox.removeAttribute('checked');
-			this.removeAttribute('checked');
-		}
-	}
-
 	disconnectedCallback() {
-		if (this.observer) {
-			this.observer.disconnect();
-			this.observer = null;
-		}
+		this.observer.disconnect();
 	}
 }
 window.customElements.define('zoo-checkbox', Checkbox);