+++ src/zoo-modules/form-modules/input-module/input.js
@@ -1,4 +1,4 @@
-import FormElement from '../common/FormElement.js';
+import { FormElement } from '../common/FormElement.js';
 
 /**
  * @injectHTML
@@ -6,13 +6,9 @@ import FormElement from '../common/FormElement.js';
 export class Input extends FormElement {
 	constructor() {
 		super();
-	}
-
-	connectedCallback() {
-		const inputSlot = this.shadowRoot.querySelector('slot[name="input"]');
-		inputSlot.addEventListener('slotchange', () => {
-			let input = inputSlot.assignedElements()[0];
-			this.registerElementForValidation(input);
+		this.shadowRoot.querySelector('slot[name="input"]').addEventListener('slotchange', e => {
+			let input = [...e.target.assignedElements()].find(el => el.tagName === 'INPUT');
+			input && this.registerElementForValidation(input);
 		});
 	}
 }