+++ src/zoo-modules/form-modules/radio-module/radio.js
@@ -1,4 +1,4 @@
-import FormElement from '../common/FormElement.js';
+import { FormElement } from '../common/FormElement.js';
 
 /**
  * @injectHTML
@@ -6,17 +6,8 @@ import FormElement from '../common/FormElement.js';
 export class Radio extends FormElement {
 	constructor() {
 		super();
-	}
-
-	connectedCallback() {
-		const radioInputSlot = this.shadowRoot.querySelector('.radio-group slot');
-		radioInputSlot.addEventListener('slotchange', () => {
-			let slottedElements = radioInputSlot.assignedElements();
-			slottedElements.forEach(e => {
-				if (e.tagName === 'INPUT') {
-					this.registerElementForValidation(e);
-				}
-			});
+		this.shadowRoot.querySelector('.radio-group slot').addEventListener('slotchange', e => {
+			e.target.assignedElements().forEach(e => e.tagName === 'INPUT' && this.registerElementForValidation(e));
 		});
 	}
 }