+++ src/zoo-modules/form-modules/searchable-select-module/searchable-select.js
@@ -1,4 +1,4 @@
-import FormElement from '../common/FormElement.js';
+import { FormElement } from '../common/FormElement.js';
 
 /**
  * @injectHTML
@@ -6,55 +6,50 @@ import FormElement from '../common/FormElement.js';
 export class SearchableSelect extends FormElement {
 	constructor() {
 		super();
-	}
-	// TODO in v9 drop nested default input and force user to define label for both select and input, while showing only legend
-	static get observedAttributes() {
-		return ['invalid', 'placeholder', 'closeicontitle'];
-	}
-	handlePlaceholder(newVal) {
-		const input = this.shadowRoot.querySelector('input');
-		if (input && newVal) input.placeholder = newVal;
-		this.inputPlaceholderFallback = newVal;
-	}
-
-	mutationCallback(mutationsList) {
-		for (let mutation of mutationsList) {
-			if (mutation.type === 'attributes' && mutation.attributeName == 'disabled') {
+		this.observer = new MutationObserver(mutationsList => {
+			for (let mutation of mutationsList) {
 				this.input.disabled = mutation.target.disabled;
 			}
-		}
-	}
-
-	connectedCallback() {
+		});
+		// TODO remove in v9
 		this.input = this.shadowRoot.querySelector('input');
 		this.input.addEventListener('input', () => this.handleSearchChange());
+		//
 		this.shadowRoot.querySelector('.cross').addEventListener('click', () => {
 			this.select.value = null;
 			this.select.dispatchEvent(new Event('change', { bubbles: true, cancelable: false }));
 		});
-		this.observer = new MutationObserver(this.mutationCallback.bind(this));
-		const selectSlot = this.shadowRoot.querySelector('slot[name="select"]');
-		selectSlot.addEventListener('slotchange', () => {
-			this.select = selectSlot.assignedElements()[0];
+		
+		this.shadowRoot.querySelector('slot[name="select"]').addEventListener('slotchange', e => {
+			this.select = [...e.target.assignedElements()].find(el => el.tagName === 'SELECT');
+			if (!this.select) return;
 			this.registerElementForValidation(this.select);
 			this.select.addEventListener('change', () => {
 				this.handleOptionChange();
 				this.valueChange();
 			});
 			this.select.size = 4;
-			this.observer.disconnect();
-			this.observer.observe(this.select, { attributes: true, childList: false, subtree: false });
+			this.observer.observe(this.select, { attributes: true, attributeFilter: ['disabled'] });
 			this.valueChange();
 			this.slotChange();
 		});
 
-		const inputSlot = this.shadowRoot.querySelector('slot[name="input"]');
-		inputSlot.addEventListener('slotchange', () => {
-			this.input = inputSlot.assignedElements()[0];
+		this.shadowRoot.querySelector('slot[name="input"]').addEventListener('slotchange', e => {
+			this.input = [...e.target.assignedElements()].find(el => el.tagName === 'INPUT');
+			if (!this.input) return;
 			this.inputPlaceholderFallback = this.input.placeholder;
 			this.input.addEventListener('input', () => this.handleSearchChange());
 			this.slotChange();
 		});
+	}
+	// TODO in v9 drop nested default input and force user to define label for both select and input, while showing only legend
+	static get observedAttributes() {
+		return ['placeholder', 'closeicontitle'];
+	}
+	handlePlaceholder(newVal) {
+		const input = this.shadowRoot.querySelector('input');
+		if (input && newVal) input.placeholder = newVal;
+		this.inputPlaceholderFallback = newVal;
 	}
 
 	slotChange() {
@@ -65,55 +60,47 @@ export class SearchableSelect extends FormElement {
 	}
 
 	valueChange() {
-		this.select.value ? this.setAttribute('valueselected', '') : this.removeAttribute('valueselected');
+		this.select.value ? this.setAttribute('value-selected', '') : this.removeAttribute('value-selected');
 	}
 
 	attributeChangedCallback(attrName, oldVal, newVal) {
-		if (SearchableSelect.observedAttributes.includes(attrName)) {
-			if (attrName == 'placeholder') {
-				this.handlePlaceholder(newVal);
-			} else if (attrName === 'invalid') {
-				const input = this.shadowRoot.querySelector('zoo-input');
-				if (input) {
-					if (this.hasAttribute('invalid')) {
-						input.setAttribute('invalid', '');
-					} else {
-						input.removeAttribute('invalid');
-					}
-				}
-			} else if (attrName === 'closeicontitle') {
-				this.shadowRoot.querySelector('zoo-cross-icon').setAttribute('title', newVal);
-			}
+		if (attrName == 'placeholder') {
+			this.handlePlaceholder(newVal);
+		} else if (attrName === 'closeicontitle') {
+			this.shadowRoot.querySelector('zoo-cross-icon').setAttribute('title', newVal);
 		}
 	}
 
 	handleSearchChange() {
 		const inputVal = this.input.value.toLowerCase();
-		const options = this.select.querySelectorAll('option');
-		for (const option of options) {
+		this.select.querySelectorAll('option').forEach(option => {
 			if (option.text.toLowerCase().indexOf(inputVal) > -1) option.style.display = 'block';
 			else option.style.display = 'none';
-		}
+		});
 	}
 
 	handleOptionChange() {
-		let inputValString = '';
-		for (const selectedOpts of this.select.selectedOptions) {
-			inputValString += selectedOpts.text + ', \n';
-		}
-		inputValString = inputValString.substr(0, inputValString.length - 3);
-		const showTooltip = inputValString && inputValString.length > 0;
-		this.input.placeholder = showTooltip ? inputValString : this.inputPlaceholderFallback;
-		if (showTooltip) {
+		let inputValString = [...this.select.selectedOptions].map(o => o.text).join(', \n');
+		this.input.placeholder = inputValString || this.inputPlaceholderFallback;
+		if (inputValString) {
 			this.input.value = null;
-			this.tooltip = this.tooltip || document.createElement('zoo-tooltip');
-			this.tooltip.slot = 'input';
-			this.tooltip.setAttribute('position', 'right');
-			this.tooltip.setAttribute('text', inputValString);
+			this.tooltip = this.tooltip || this.createTooltip();
+			this.tooltip.innerHTML = inputValString;
 			this.shadowRoot.querySelector('zoo-input').appendChild(this.tooltip);
 		} else if (this.tooltip) {
 			this.tooltip.remove();
 		}
+	}
+
+	createTooltip() {
+		const tooltip = document.createElement('zoo-tooltip');
+		tooltip.slot = 'additional';
+		tooltip.setAttribute('position', 'right');
+		return tooltip;
+	}
+
+	disconnectedCallback() {
+		this.observer.disconnect();
 	}
 }
 window.customElements.define('zoo-searchable-select', SearchableSelect);