+++ src/zoo-modules/form-modules/select-module/select.js
@@ -1,4 +1,4 @@
-import FormElement from '../common/FormElement.js';
+import { FormElement } from '../common/FormElement.js';
 
 /**
  * @injectHTML
@@ -6,41 +6,24 @@ import FormElement from '../common/FormElement.js';
 export class Select extends FormElement {
 	constructor() {
 		super();
-	}
-
-	mutationCallback(mutationsList) {
-		for(let mutation of mutationsList) {
-			if (mutation.type === 'attributes') {
+		this.observer = new MutationObserver(mutationsList => {
+			for(let mutation of mutationsList) {
 				const attr = mutation.attributeName;
-				if (attr == 'disabled' || attr == 'multiple') {
-					if (mutation.target[attr]) {
-						this.setAttribute(attr, '');
-					} else {
-						this.removeAttribute(attr);
-					}
-				}
+				mutation.target[attr] ? this.setAttribute(attr, '') : this.removeAttribute(attr);
 			}
-		}
-	}
-
-	connectedCallback() {
-		const selectSlot = this.shadowRoot.querySelector('slot[name="select"]');
-		selectSlot.addEventListener('slotchange', () => {
-			this.observer = this.observer || new MutationObserver(this.mutationCallback.bind(this));
-			let select = selectSlot.assignedElements()[0];
-			if (select.hasAttribute('multiple')) this.setAttribute('multiple', '');
-			if (select.hasAttribute('disabled')) this.setAttribute('disabled', '');
+		});
+		this.shadowRoot.querySelector('slot[name="select"]').addEventListener('slotchange', e => {
+			let select = [...e.target.assignedElements()].find(el => el.tagName === 'SELECT');
+			if (!select) return;
+			if (select.multiple) this.setAttribute('multiple', '');
+			if (select.disabled) this.setAttribute('disabled', '');
 			this.registerElementForValidation(select);
-			this.observer.disconnect();
-			this.observer.observe(select, { attributes: true, childList: false, subtree: false });
+			this.observer.observe(select, { attributes: true, attributeFilter: ['disabled', 'multiple'] });
 		});
 	}
 
 	disconnectedCallback() {
-		if (this.observer) {
-			this.observer.disconnect();
-			this.observer = null;
-		}
+		this.observer.disconnect();
 	}
 }
 window.customElements.define('zoo-select', Select);