+++ src/zoo-modules/form-modules/toggle-switch-module/toggle-switch.js
@@ -1,4 +1,4 @@
-import FormElement from '../common/FormElement.js';
+import { FormElement } from '../common/FormElement.js';
 
 /**
  * @injectHTML
@@ -6,16 +6,10 @@ import FormElement from '../common/FormElement.js';
 export class ToggleSwitch extends FormElement {
 	constructor() {
 		super();
-	}
-
-	connectedCallback() {
-		const inputSlot = this.shadowRoot.querySelector('slot[name="input"]');
-		inputSlot.addEventListener('slotchange', () => {
-			let input = inputSlot.assignedElements()[0];
+		this.shadowRoot.querySelector('slot[name="input"]').addEventListener('slotchange', e => {
+			const input = [...e.target.assignedElements()].find(el => el.tagName === 'INPUT');
+			if (!input) return;
 			this.registerElementForValidation(input);
-			input.addEventListener('change', e => {
-				e.target.checked ? e.target.setAttribute('checked', '') : e.target.removeAttribute('checked');
-			});
 		});
 	}
 }