+++ src/zoo-modules/misc-modules/collapsable-list-item-module/collapsable-list-item.js
@@ -4,17 +4,14 @@
 export class CollapsableListItem extends HTMLElement {
 	constructor() {
 		super();
-	}
-
-	connectedCallback() {
-		const details = this.shadowRoot.querySelector('details');
-		details.addEventListener('toggle', e => {
+		this.details = this.shadowRoot.querySelector('details');
+		this.details.addEventListener('toggle', e => {
 			this.shadowRoot.host.dispatchEvent(new CustomEvent('toggle', {detail: e.target.open, composed: true}));
 		});
 	}
 
 	close() {
-		this.shadowRoot.querySelector('details').open = false;
+		this.details.open = false;
 	}
 }
 window.customElements.define('zoo-collapsable-list-item', CollapsableListItem);