+++ src/zoo-modules/misc-modules/toast-module/toast.js
@@ -5,11 +5,13 @@ export class Toast extends HTMLElement {
 	constructor() {
 		super();
 	}
+	
 	connectedCallback() {
 		this.hidden = true;
 		this.timeout = this.getAttribute('timeout') || 3;
 		this.setAttribute('role', 'alert');
 	}
+	
 	show() {
 		if (!this.hidden) return;
 		this.style.display = 'block';